Helper function for get_metrics

This is a helper function that computes the desired metrics from a data.frame of the top terms for each source node in a network.

Usage

get_network_metrics(
  full_terms,
  network_size,
  organism,
  gene_id,
  get_sum,
  get_percent,
  get_mean,
  get_median,
  get_annotation_overlap,
  get_size,
  penalty,
  fdr_threshold,
  go_ann = NULL,
  go_reg = NULL
)

Arguments

full_terms

a data.frame containing all the summary results for a network. Use get_terms to obtain.

network_size

the number of source nodes in the network

organism

a string specifying the organism that the data is from, e.g. "hsapiens" or "scerevisiae"

gene_id

the naming system used for the input genes - see options with WebGestaltR::listIdType() and see webgestalt.org for examples of each type

get_sum

boolean whether to get the 'sum' metric, which is the sum of the negative log base 10 of the p-value for the top term of each source node minus 'penalty' times the total number of source nodes.

get_percent

boolean whether to get the 'percent' metric, which is the percent of source nodes with at least one term with a FDR below the 'fdr_threshold'

get_mean

boolean whether to get the 'mean' metric, which is the mean negative log base 10 of the p-value for the top term of each source node regardless of significance

get_median

boolean whether to get the 'median' metric, which is the median negative log base 10 of the p-value for the top term of each source node regardless of significance

get_annotation_overlap

boolean whether to get the 'annotation_overlap' metric, which is the percent of source nodes that are annotated to at least one of the 16 GO terms for which their target genes are most enriched

get_size

boolean whether to get the 'size' metric, which is the number of source nodes in the network subset that have more than one target gene with annotations. This number is used in the calculation of all other metrics.

penalty

the penalty applied to the 'sum' metric for each source node in the network

fdr_threshold

the FDR threshold for a gene set term to be considered significantly over-represented for the purposes of calculating the 'percent' metric

go_ann

a data.frame of annotations of genes to GO terms. Obtain with WebGestaltR::loadGeneSet.

go_reg

a data.frame of the regulatory relationships between GO terms. Obtain with ontologyIndex::get_ontology.

Value

a list of metric values